Move output functions to own file
